planventure-api/requirements.txt
# Core dependencies
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
flask-jwt-extended==4.5.3
python-dotenv==1.0.0
bcrypt==4.0.1
PyJWT==2.8.0